cart merge works on cart link click
