+++ client/components/navbar.js
@@ -3,13 +3,16 @@ import PropTypes from 'prop-types'
 import {connect} from 'react-redux'
 import {Link} from 'react-router-dom'
 import {logout} from '../store'
+import axios from 'axios'
 
-const Navbar = ({handleClick, isLoggedIn}) => (
+const Navbar = ({handleClick, isLoggedIn, handleCart, user}) => (
   <div>
     <h1>Vintage Vidya Games!</h1>
     <nav>
       <Link to="/home">Home</Link>
-      <Link to="/cart">Cart</Link>
+      <Link to="/cart" onClick={() => handleCart(user)}>
+        Cart
+      </Link>
       {isLoggedIn ? (
         <div>
           {/* The navbar will show these links after you log in */}
@@ -34,7 +37,8 @@ const Navbar = ({handleClick, isLoggedIn}) => (
  */
 const mapState = state => {
   return {
-    isLoggedIn: !!state.user.id
+    isLoggedIn: !!state.user.id,
+    user: state.user
   }
 }
 
@@ -43,6 +47,44 @@ const mapDispatch = dispatch => {
     handleClick() {
       dispatch(logout())
       localStorage.clear()
+    },
+    async handleCart(user) {
+      if (user.id) {
+        const {data} = await axios.get(`/api/orders/${user.id}`)
+        if (data.length) {
+          if (localStorage.cart) {
+            const curCart = JSON.parse(localStorage.cart)
+            data.forEach(async curItem => {
+              if (curCart[curItem.itemId]) {
+                if (curCart[curItem.itemId] !== curItem.quantity) {
+                  const curGuestCartQuantity = curCart[curItem.itemId]
+                  curCart[curItem.itemId] =
+                    curCart[curItem.itemId] + curItem.quantity
+                  const updateCurItemQuantity = await axios.put(
+                    `/api/orders/edit/${user.id}`,
+                    {
+                      item: {
+                        id: curItem.itemId,
+                        price: curItem.price,
+                        quantity: curGuestCartQuantity
+                      }
+                    }
+                  )
+                }
+              } else {
+                curCart[curItem.itemId] = curItem.quantity
+              }
+            })
+            localStorage.setItem('cart', JSON.stringify(curCart))
+          } else {
+            const retrievedCart = data.reduce((acc, curItem) => {
+              acc[curItem.itemId] = curItem.quantity
+              return acc
+            }, {})
+            localStorage.setItem('cart', JSON.stringify(retrievedCart))
+          }
+        }
+      }
     }
   }
 }
@@ -54,5 +96,7 @@ export default connect(mapState, mapDispatch)(Navbar)
  */
 Navbar.propTypes = {
   handleClick: PropTypes.func.isRequired,
-  isLoggedIn: PropTypes.bool.isRequired
+  isLoggedIn: PropTypes.bool.isRequired,
+  handleCart: PropTypes.func.isRequired,
+  user: PropTypes.object.isRequired
 }
